# xadminapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "adminapi",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin_user/create_admin_user": {
            "post": {
                "tags": [
                    "后台用户"
                ],
                "summary": "新增管理员账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.CreateAdminUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/admin_user/delete_admin_user": {
            "delete": {
                "tags": [
                    "后台用户"
                ],
                "summary": "删除管理员账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.DeleteAdminUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/admin_user/get_admin_user": {
            "get": {
                "tags": [
                    "后台用户"
                ],
                "summary": "获取管理员列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "账号",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "渠道商",
                        "name": "channel_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "运营商",
                        "name": "seller_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.XAdminUser"
                            }
                        }
                    }
                }
            }
        },
        "/admin_user/set_login_googlesecret": {
            "post": {
                "tags": [
                    "后台用户"
                ],
                "summary": "设置登录验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.SetLoginGoogleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/service_admin.SetLoginGoogleRes"
                        }
                    }
                }
            }
        },
        "/admin_user/set_opt_googlesecret": {
            "post": {
                "tags": [
                    "后台用户"
                ],
                "summary": "设置操作验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.SetOptGoogleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/service_admin.SetOptGoogleRes"
                        }
                    }
                }
            }
        },
        "/admin_user/update_admin_user": {
            "patch": {
                "tags": [
                    "后台用户"
                ],
                "summary": "更新管理员账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.UpdateAdminUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/admin_user/user_login": {
            "post": {
                "tags": [
                    "后台登录"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.AdminUserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/service_admin.AdminUserLoginRes"
                        }
                    }
                }
            }
        },
        "/admin_user/user_logout": {
            "post": {
                "tags": [
                    "后台登录"
                ],
                "summary": "管理员退出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.XAdminUser": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "自增Id",
                    "type": "integer"
                },
                "login_count": {
                    "description": "登录次数",
                    "type": "integer"
                },
                "login_ip": {
                    "description": "最后登录Ip",
                    "type": "string"
                },
                "login_time": {
                    "description": "最后登录时间",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "role_name": {
                    "description": "角色",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "service_admin.AdminUserLoginReq": {
            "type": "object",
            "required": [
                "account",
                "password"
            ],
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "service_admin.AdminUserLoginRes": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "auth_data": {
                    "description": "权限数据",
                    "type": "string"
                },
                "login_count": {
                    "description": "登录次数",
                    "type": "integer"
                },
                "login_ip": {
                    "description": "登录Ip",
                    "type": "string"
                },
                "login_time": {
                    "description": "登录时间",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "token": {
                    "description": "token",
                    "type": "string"
                },
                "utc_offset": {
                    "description": "当地时区与utc的偏移量",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.CreateAdminUserReq": {
            "type": "object",
            "required": [
                "account",
                "password",
                "role_name",
                "seller_id",
                "state"
            ],
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "password": {
                    "description": "登录密码",
                    "type": "string"
                },
                "role_name": {
                    "description": "角色",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.DeleteAdminUserReq": {
            "type": "object",
            "required": [
                "account",
                "seller_id"
            ],
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.SetLoginGoogleReq": {
            "type": "object",
            "required": [
                "account",
                "seller_id"
            ],
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                }
            }
        },
        "service_admin.SetLoginGoogleRes": {
            "type": "object",
            "properties": {
                "google_secret": {
                    "description": "登录验证码",
                    "type": "string"
                }
            }
        },
        "service_admin.SetOptGoogleReq": {
            "type": "object",
            "required": [
                "account",
                "seller_id"
            ],
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                }
            }
        },
        "service_admin.SetOptGoogleRes": {
            "type": "object",
            "properties": {
                "google_secret": {
                    "description": "操作验证码",
                    "type": "string"
                }
            }
        },
...
        "service_admin.UpdateAdminUserReq": {
            "type": "object",
            "required": [
                "account",
                "seller_id"
            ],
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string"
                },
                "channel_id": {
                    "description": "渠道商",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "password": {
                    "description": "登录密码",
                    "type": "string"
                },
                "role_name": {
                    "description": "角色",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}